+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 - feat(sourcemaps): Add bundler selection prompt (#304)
 - feat(sourcemaps): Add Login and Project Selection flow (#300)
 - feat(sourcemaps): Add setup flow for Vite (#308)
+- feat(sourcemaps): Add setup flow for sentry-cli (#314)
 - feat(sourcemaps): Add Sourcemaps as selectable integration (#302)
 - feat(sourcemaps): Create `.env.sentry-build-plugin` instead of `.sentryclirc` to set auth token (#313)
 - feat: Add empty sourcemaps wizard (#295)